fixed #141 
fixed #77 
fixed #144 
fixed #124 
fixed #148 
fixed #147 
fixed #112
fixed #74

+++ src-tauri/src/commands/db/periods.rs
@@ -1,5 +1,6 @@
 use error_stack::{IntoReport, ResultExt};
 use sea_orm::DatabaseConnection;
+use tokio::sync::Mutex;
 
 use entity::prelude::Period;
 
@@ -7,6 +8,8 @@ use crate::{
 	commands::utils::{CommandError, Delete, LogAndString},
 	db::periods::{create_period, delete_period, edit_period, get_periods},
 };
+use crate::config::main::Config;
+use crate::utils::StrError;
 
 #[tauri::command]
 pub async fn get_periods_js(connection: tauri::State<'_, DatabaseConnection>) -> Result<String, String> {
@@ -65,7 +68,7 @@ pub async fn edit_period_js(connection: tauri::State<'_, DatabaseConnection>, js
 }
 
 #[tauri::command]
-pub async fn delete_period_js(connection: tauri::State<'_, DatabaseConnection>, json: String) -> Result<(), String> {
+pub async fn delete_period_js(connection: tauri::State<'_, DatabaseConnection>, config: tauri::State<'_, Mutex<Config>>, json: String) -> Result<(), String> {
 	log::debug!("json delete period: {}", json);
 	
 	let delete: Delete = serde_json::from_str(&json)
@@ -74,6 +77,18 @@ pub async fn delete_period_js(connection: tauri::State<'_, DatabaseConnection>, 
 			.attach_printable_lazy(|| format!("json: {}", json))
 			.change_context(CommandError)
 			.log_and_to_string()?;
+	
+	let grade_modal_defaults = {
+		let mutex = config.lock().await;
+		mutex.get().grade_modal_defaults.clone()
+	};
+	
+	if grade_modal_defaults.period_default == Some(delete.id) {
+		return Err(StrError("Cannot delete default period".to_string()))
+				.into_report()
+				.change_context(CommandError)
+				.log_and_to_string();
+	}
 	
 	delete_period(&connection, delete.id)
 			.await
